+++ mutwo/music_parameters/__init__.py
@@ -2,6 +2,7 @@ from . import configurations
 from . import constants
 from . import abc
 
+from .ambituses import *
 from .commas import *
 from .pitches import *
 from .volumes import *
@@ -9,4 +10,4 @@ from .notation_indicators import *
 from .playing_indicators import *
 
 # Force flat structure
-del commas, pitches, volumes, notation_indicators, playing_indicators
+del ambituses, commas, pitches, volumes, notation_indicators, playing_indicators